+++ Source/Workers/createPolylineGeometry.js
@@ -1,9 +1,16 @@
 /*global define*/
 define([
+        '../Core/Ellipsoid',
         '../Core/PolylineGeometry'
     ], function(
+        Ellipsoid,
         PolylineGeometry) {
     "use strict";
 
-    return PolylineGeometry.createGeometry;
+    function createPolylineGeometry(polylineGeometry) {
+        polylineGeometry._ellipsoid = Ellipsoid.clone(polylineGeometry._ellipsoid);
+        return PolylineGeometry.createGeometry(polylineGeometry);
+    }
+
+    return createPolylineGeometry;
 });