+++ CHANGES.md
@@ -67,6 +67,7 @@ Change Log
     * `Rectangle.getNortheast` -> `Rectangle.northeast`
     * `Rectangle.getCenter` -> `Rectangle.center`
     * `CullingVolume.getVisibility` -> `CullingVolume.computeVisibility`
+  * `SimplePolylineGeometry` and `PolylineGeometry` positions curve to follow the ellipsoid surface by default. To disable this behavior, set the option `followSurface=false`.
 * Added northUpEast transform to help support display of glTF models because Y is their up axis.
 * Cesium can now render an unlimited number of imagery layers, no matter how few texture units are supported by the hardware.
 * Added `Primitive.ready`.